+++ content/about/we.md
@@ -1,0 +1,9 @@
+# 关于 Minecraft集体宣传组织(MSCPO)
+
+Minecraft集体宣传组织(MSCPO)是一个致力于我的世界服务器宣传和技术交流的新兴组织（×）服主同好群（√）
+
+> 委员会状态： 已经重组
+
+## 服务器列表项目信息
+
+![Alt](https://repobeats.axiom.co/api/embed/40f1d6f2bafcf731a005ddad0f4a0178ce8e6c46.svg "Repobeats analytics image")
