+++ pages/about/index.vue
@@ -1,172 +1,160 @@
 <script setup lang="ts">
-import commit_hash from '@/assets/commit-hash.svg'
-import Logo from '@/assets/logo.svg'
-const title = useState<string>('pageTitle')
-title.value = '关于 MineShowcase'
-
-useHead({
-    title: title,
-})
-const runtimeConfig = useRuntimeConfig()
-const commitHash = runtimeConfig.public.commitHash
-    ? runtimeConfig.public.commitHash.slice(0, 8)
-    : 'dev build'
-const githubCommitUrl = computed(() =>
-    runtimeConfig.public.commitHash
-        ? `https://github.com/MSCPO/MineShowcase/commit/${runtimeConfig.public.commitHash}`
-        : 'https://github.com/MSCPO/MineShowcase',
+const { data: testMd, error } = await useAsyncData(() =>
+    queryCollection('content').path('/about/we').first(),
 )
+
+if (error.value) {
+    console.error('Failed to load markdown content:', error.value)
+}
+
+useSeoMeta({
+    title: testMd.value?.title || 'Default Title',
+    description: testMd.value?.description || 'Default Description',
+})
 </script>
 
 <template>
-    <div class="about-box">
-        <div class="logo-box">
-            <img class="logo-img" :src="Logo" />
-            <div class="logo-title">
-                <h1>MineShowcase</h1>
-                <div class="ver">
-                    Commit: (
-                    <img :src="commit_hash" alt="commit 图标" />
-                    <a
-                        :href="githubCommitUrl"
-                        target="_blank"
-                        rel="noopener noreferrer"
-                    >
-                        {{ commitHash }}
-                    </a>
-                    )
-                </div>
-                <div class="icons">
-                    <a-tooltip title="项目 Github 仓库">
-                        <a
-                            href="https://github.com/MSCPO/MineShowcase"
-                            class="icon"
-                        >
-                            <GithubOutlined />
-                        </a>
-                    </a-tooltip>
-
-                    <a-tooltip title="玩家交流群">
-                        <a href="https://qm.qq.com/q/SHcbys4Scw" class="icon">
-                            <QqOutlined />
-                        </a>
-                    </a-tooltip>
-
-                    <!-- <ClientOnly>
-                        <a-tooltip
-                            v-if="$pwa?.isPWAInstalled"
-                            title="应用已安装"
-                        >
-                            <div class="icon">
-                                <AppstoreOutlined />
-                            </div>
-                        </a-tooltip>
-                        <a-tooltip v-else title="安装此应用">
-                            <div @click="$pwa?.install" class="icon">
-                                <AppstoreAddOutlined />
-                            </div>
-                        </a-tooltip>
-                    </ClientOnly> -->
-                </div>
-            </div>
-        </div>
-
-        <a-card title="关于 MineShowcase">
-            <p>
-                MineShowcase 是一个以 MSCPO/ServerAPI 作为后端的 MSCPO
-                开源宣传平台
-            </p>
-            <p>此项目基于 Vue 3 语言，以及 Nuxt 3 框架</p>
-            <p>MineShowcase © 2025 MSCPO 保留所有权利。</p>
-        </a-card>
+    <div class="markdown-container">
+        <ContentRenderer v-if="testMd" :value="testMd" class="content" />
+        <div v-else>Home not found</div>
     </div>
 </template>
 
 <style scoped lang="less">
-@import '@/assets/css/variables.less';
-.about-box {
-    text-align: center;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-items: center;
-    padding: 1rem;
-    .logo-box {
-        display: flex;
-        box-sizing: border-box;
-        padding: 4rem;
-        @media screen and (max-width: 768px) {
-            padding: 0;
-            flex-direction: column;
-            align-items: center;
-            padding-block-end: 1rem;
+.markdown-container {
+    max-width: 800px;
+    margin: 0 auto;
+    padding: 2rem 1rem;
+    .content {
+        display: block !important;
+        :deep(*) {
+            // 通用样式
+            line-height: 1.6;
+            color: #374151;
+            max-width: 100%;
         }
-        .logo-img {
-            width: 8rem;
-        }
-        .logo-title {
-            color: @text-color-light;
-            display: flex;
-            line-height: 1rem;
-            flex-direction: column;
-            padding-inline: 1rem;
-            @media (prefers-color-scheme: dark) {
-                color: @text-color-dark;
-            }
-            .ver {
-                font-size: 0.8rem;
-                color: @text-color-light;
-                display: flex;
-                align-items: center;
-                gap: 0.5rem;
-                @media (prefers-color-scheme: dark) {
-                    color: @text-color-dark;
-                }
-                img {
-                    width: 1rem;
-                    height: 1rem;
-                    filter: invert(1);
-                }
-            }
-            .ver a {
-                color: inherit;
-                text-decoration: none;
-                transition: color 0.3s ease;
 
-                &:hover {
-                    color: @hover-primary;
-                    text-decoration: underline;
-                }
+        // 标题样式
+        :deep(h1) {
+            font-size: 2.5rem;
+            border-bottom: 2px solid #e5e7eb;
+            padding-bottom: 0.5rem;
+            margin: 2rem 0 1.5rem;
+        }
+
+        :deep(h2) {
+            font-size: 2rem;
+            margin: 1.8rem 0 1.2rem;
+            color: #1f2937;
+        }
+
+        :deep(h3) {
+            font-size: 1.5rem;
+            margin: 1.5rem 0 1rem;
+            color: #374151;
+        }
+
+        // 段落和文字
+        :deep(p) {
+            margin: 1rem 0;
+            font-size: 1.1rem;
+        }
+
+        :deep(a) {
+            text-decoration: none;
+            &:hover {
+                text-decoration: underline;
             }
-            .icons {
-                width: 100%;
-                padding-top: 1rem;
-                display: flex;
-                justify-content: center;
-                gap: 1rem;
-                .icon {
-                    color: @text-color-light;
-                    @media (prefers-color-scheme: dark) {
-                        color: @text-color-dark;
-                    }
-                    width: 32px;
-                    height: 32px;
-                    border-radius: 1rem;
-                    word-break: break-all;
-                    text-decoration: none;
-                    display: flex;
-                    justify-content: center;
-                    cursor: pointer;
-                    align-items: center;
-                    font-size: 1.5rem;
-                    text-align: center;
-                    transition: all 0.3s ease;
-                    &:hover {
-                        background: @hover-primary;
-                        transform: scale(1.2);
-                    }
-                }
+        }
+
+        :deep(strong) {
+            color: #1f2937;
+        }
+
+        :deep(em) {
+            color: #4b5563;
+        }
+
+        // 列表
+        :deep(ul),
+        :deep(ol) {
+            margin: 1rem 0;
+            padding-left: 2rem;
+        }
+
+        :deep(li) {
+            margin: 0.5rem 0;
+            &::marker {
+                color: #6b7280;
             }
+        }
+
+        // 代码块
+        :deep(pre) {
+            background-color: #f3f4f6;
+            border-radius: 8px;
+            padding: 1.5rem;
+            margin: 1.5rem 0;
+            overflow-x: auto;
+            code {
+                font-family:
+                    'SFMono-Regular', Consolas, 'Liberation Mono', Menlo,
+                    monospace;
+                font-size: 0.9rem;
+                color: #1f2937;
+            }
+        }
+
+        :deep(code:not(pre code)) {
+            background-color: #f3f4f6;
+            padding: 0.2em 0.4em;
+            border-radius: 4px;
+            font-size: 0.9em;
+        }
+
+        // 表格
+        :deep(table) {
+            width: 100%;
+            margin: 1.5rem 0;
+            border-collapse: collapse;
+            box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+
+            th,
+            td {
+                padding: 0.75rem;
+                border: 1px solid #e5e7eb;
+            }
+
+            th {
+                background-color: #f9fafb;
+                font-weight: 600;
+            }
+
+            tr:nth-child(even) {
+                background-color: #f9fafb;
+            }
+        }
+
+        // 引用
+        :deep(blockquote) {
+            border-left: 4px solid #e5e7eb;
+            margin: 1.5rem 0;
+            padding: 1rem 1.5rem;
+            background-color: #f9fafb;
+            color: #4b5563;
+            p {
+                margin: 0;
+            }
+        }
+
+        // 图片
+        :deep(img) {
+            max-width: 100%;
+            height: auto;
+            margin: 1.5rem auto;
+            display: block;
+            border-radius: 8px;
+            box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
         }
     }
 }
